feat: support fractional weight (double) for sets
- Set: change weight field from int to double
- SetModel: Q_PROPERTY weight changed to double
- SetSerializer: use toDouble() for weight deserialization
- SetRepositoryDb: bind weight as double
- ExerciseModel/Exercise: setsToString formats weight as double
- WorkoutJson: serialize/deserialize weight as double
- RestDialog: minor formatting fix
- Tests: add fractional weight tests (25.5, 27.5, 0.5kg steps)
  and update all weight assertions to EXPECT_DOUBLE_EQ

+++ src/modules/workout/domain/entities/set.cpp
@@ -2,7 +2,7 @@
 
 Set::Set() = default;
 
-Set::Set(int repetitions, int weight)
+Set::Set(int repetitions, double weight)
     : m_repetitions(repetitions), m_weight(weight)
 {
     validate();
@@ -11,13 +11,13 @@ Set::Set(int repetitions, int weight)
 int Set::id() const { return m_id; }
 int Set::exerciseId() const { return m_exerciseId; }
 int Set::repetitions() const { return m_repetitions; }
-int Set::weight() const { return m_weight; }
+double Set::weight() const { return m_weight; }
 bool Set::completed() const { return m_completed; }
 
 void Set::setId(int id) { m_id = id; }
 void Set::setExerciseId(int exerciseId) { m_exerciseId = exerciseId; }
 void Set::setRepetitions(int repetitions) { m_repetitions = repetitions; }
-void Set::setWeight(int weight) { m_weight = weight; }
+void Set::setWeight(double weight) { m_weight = weight; }
 void Set::setCompleted(bool completed) { m_completed = completed; }
 
 void Set::validate() const
