+++ src/modules/workout/domain/entities/set.h
@@ -4,18 +4,18 @@ class Set final
 {
 public:
     Set();
-    Set(int repetitions, int weight);
+    Set(int repetitions, double weight);
 
     int id() const;
     int exerciseId() const;
     int repetitions() const;
-    int weight() const;
+    double weight() const;
     bool completed() const;
 
     void setId(int id);
     void setExerciseId(int exerciseId);
     void setRepetitions(int repetitions);
-    void setWeight(int weight);
+    void setWeight(double weight);
     void setCompleted(bool completed);
 
 private:
@@ -24,6 +24,6 @@ private:
     int m_id{-1};
     int m_exerciseId{-1};
     int m_repetitions{0};
-    int m_weight{0};
+    double m_weight{0.0};
     bool m_completed{false};
 };
